UPDATE - correct Learning Rate input

+++ main.py
@@ -10,7 +10,7 @@ import torch.nn as nn
 ###Constants
 model_select = int(input("Select your model : Enter 1 for vgg , 2 for resnet 18, 3 for mobilenet\n"))
 num_epochs = int(input("Select epochs number\n"))
-lr = int(input("Select learning rate\n"))
+lr = float(input("Select learning rate\n"))
 
 if model_select == 1:
     model_applied = models.vgg16(pretrained=True)
